Added unit tests to CPOptions; fixed __str__ for AWSPath.
Enabled a number of unit tests for CPOptions, primarily around copying from a directory and from an AWS resource.  Added some argument checking to CPOptions to check that the target is not the same as one of the sources.
Fixed a bug in AWSPath where the string representation of an AWSPath was displaying the scheme ('aws') twice.  Added a unit test to validate the fix.